+++ apps/extension/src/constants/routes/index.ts
@@ -1,0 +1,6 @@
+export const Routes = {
+  root: "/",
+  settings: "settings",
+  info: "info",
+  any: "*",
+};