+++ apps/extension/src/pages/index.tsx
@@ -1,0 +1,5 @@
+export * from "@pages/Home";
+export * from "@pages/Info";
+export * from "@pages/Settings";
+export * from "@pages/NotFound";
+export * from "@pages/Root";