+++ apps/extension/src/providers/index.tsx
@@ -1,0 +1,2 @@
+export * from "@providers/Redux";
+export * from "@providers/Style";