+++ apps/extension/CHANGELOG.md
@@ -1,5 +1,6 @@
-## [2.0.0] - 03-07-2025
+## [2.0.0] - 06-07-2025
 
+- ✨ feat: add new page(Info), content for Settings/Info page and modify imports using typescript paths
 - ✨ feat: add routing(home and settings)
 - ✨ feat: add functionality for dark/light/system theme
 - ✨ feat: add redux/redux-persist and initial state for settings